refactor(agent): deepen chat trigger input handling
